fixes to student folder

+++ fe/student/views/credit.py
@@ -3,12 +3,13 @@ from student.models import Student
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 
+# FIXME: add same logic, just students accounts can access this page
 @login_required
 def top_up_student_credit(request):
     student = Student.objects.get(user=request.user)
-
+    
     if request.method == 'POST':
-        amount = float(request.POST.get('amount'))
+        amount = int(request.POST.get('amount'))
         student.credit += amount
         student.save()
         return redirect('SuccessPage')
